requirements.txt
Flask>=2.3
numpy>=1.24
pandas>=2.0
scikit-learn>=1.2
joblib>=1.2
pyttsx3>=2.90
requests>=2.31
pyserial>=3.5
pytest>=7.4
black>=23.7
isort>=5.12
flake8>=6.1
